Extracted some logic from DriftRawData.py to DriftsInterpolator.py
